+++ src/Sentry/ITransactionContext.cs
@@ -14,4 +14,9 @@ public interface ITransactionContext : ISpanContext
     /// Whether the parent transaction of this transaction has been sampled.
     /// </summary>
     bool? IsParentSampled { get; }
+
+    /// <summary>
+    /// The source of the transaction name.
+    /// </summary>
+    TransactionNameSource NameSource { get; }
 }